+++ src/StateSmith.Cli/Create/TemplateIds.cs
@@ -4,6 +4,7 @@ public static class TemplateIds
 {
     public const string _common = "_common";
     public const string DrawIoSimple1 = "drawio-simple-1";
+    public const string DrawIoPages1 = "drawio-pages-1";
     public const string PlantUmlSimple1 = "plantuml-simple-1";
     public const string PlantUmlMinimal1 = "plantuml-minimal-1";
 }