+++ src/StateSmith.Cli/Run/SsDiagramFilter.cs
@@ -46,8 +46,9 @@ public class SsDiagramFilter
 
     public bool IsSsPlantUmlFileContents(string diagramContents)
     {
-        // Need to find `@startuml <someWord>` and `->`
-        var regex = new Regex(@"@startuml\s+\w+");
+        // Need to find `@startuml <nonSpace>` and `->`
+        // Nonspace to support `{fileNames}` in the diagram. https://github.com/StateSmith/StateSmith/issues/330
+        var regex = new Regex(@"@startuml[ \t]+\S+");
         return regex.IsMatch(diagramContents) && diagramContents.Contains("->");
         // Don't test for `[*]` as new users may forget to add it initially.
     }